[AccOpening] verify current user who is deleting object if its the same as connected one

+++ controllers/product.js
@@ -39,7 +39,19 @@ exports.updateProduct = (req, res, next) => {
 };
 
 exports.deleteProduct = (req, res) => {
-  Product.deleteOne({ _id: req.params.id })
-    .then((data) => res.json({ data, message: "deleted successfully" }))
-    .catch((error) => res.json({ error }));
+  Product.findOne({ _id: req.params.id }).then((product) => {
+    if (!product) {
+      return res.status(404).json({
+        error: new Error("No product found"),
+      });
+    }
+    if (product.userId !== req.auth.userId) {
+      return res.status(401).json({
+        error: new Error("Unauthorized"),
+      });
+    }
+    Product.deleteOne({ _id: req.params.id })
+      .then((data) => res.json({ data, message: "deleted successfully" }))
+      .catch((error) => res.json({ error }));
+  });
 };
